:truck: (interface): Rename file interface/CoreFont.hpp to interface/Font.hpp
